backend/requirements.txt
fastapi
uvicorn
groq
chromadb
sentence-transformers
pypdf
python-dotenv
langgraph